Rebase and move tests into the main Varuna tests module
