+++ algorithms/src/snark/varuna/mod.rs
@@ -29,6 +29,3 @@ pub use mode::*;
 
 #[cfg(test)]
 pub mod tests;
-
-#[cfg(test)]
-pub mod sage_test;
